fixed gunAngle, fixed boat movement

+++ visor-core/src/main/java/me/phoenixra/visor/core/client/data/raw/RawPoseHandler.java
@@ -17,7 +17,7 @@ public abstract class RawPoseHandler {
     protected RawController controllerRightData;
 
     @Getter
-    protected double gunAngle = 0.0D;
+    protected float gunAngle = 0.0f;
